Refactor SHA message digest tests into JUnit 5 parameterized tests
Convert SHA message digest test implementations to JUnit 5 parameterized
tests and remove duplicate provider-specific test code.

Changes:

- Convert BaseTestSHA1.java to a JUnit 5 parameterized test
- Convert BaseTestSHA224.java to a JUnit 5 parameterized test
- Convert BaseTestSHA256.java to a JUnit 5 parameterized test
- Convert BaseTestSHA384.java to a JUnit 5 parameterized test
- Convert BaseTestSHA512.java to a JUnit 5 parameterized test
- Convert BaseTestSHA512_224.java to a JUnit 5 parameterized test
- Convert BaseTestSHA512_256.java to a JUnit 5 parameterized test
- Remove provider-specific test wrapper classes
- Update TestAll.java to remove deleted test references
- Update TestMultithread.java to remove references to deleted tests
- Update TestMultithreadFIPS.java to remove references to deleted tests
