+++ src/test/java/ibm/jceplus/junit/TestMultithread.java
@@ -55,9 +55,6 @@ public class TestMultithread {
             "ibm.jceplus.junit.openjceplus.multithread.TestRSAPSSInterop3",
             "ibm.jceplus.junit.openjceplus.multithread.TestRSASignatureInteropSunRsaSign",
             "ibm.jceplus.junit.openjceplus.multithread.TestSHA256Clone_SharedMD",
-            "ibm.jceplus.junit.openjceplus.multithread.TestSHA512",
-            "ibm.jceplus.junit.openjceplus.multithread.TestSHA512_224",
-            "ibm.jceplus.junit.openjceplus.multithread.TestSHA512_256",
             "ibm.jceplus.junit.openjceplus.multithread.TestXDH",
             "ibm.jceplus.junit.openjceplus.multithread.TestXDHKeyImport",
             "ibm.jceplus.junit.openjceplus.multithread.TestXDHKeyPairGenerator",
