+++ src/test/java/ibm/jceplus/junit/openjceplus/TestAll.java
@@ -95,13 +95,6 @@ import org.junit.platform.suite.api.Suite;
     TestRSASignatureInteropSunRsaSign.class,
     TestRSATypeCheckDefault.class,
     TestRSATypeCheckEnabled.class,
-    TestSHA1.class,
-    TestSHA224.class,
-    TestSHA256.class,
-    TestSHA384.class,
-    TestSHA512_224.class,
-    TestSHA512_256.class,
-    TestSHA512.class,
     TestXDH.class,
     TestXDHInterop.class,
     TestXDHInteropBC.class,
